Removed id_translation from sprk queries and directly copied results. Directly copying rippled through other files, as sprk needs 64bit ids for some reason

+++ src/embeddingLib/include/spacialQuery/SpacialIndex.hpp
@@ -11,13 +11,13 @@ class SpatialIndex {
     virtual ~SpatialIndex() = default;
 
     // Query for nearest neighbors (point-based query)
-    virtual size_t query_nearest(CVecRef point, unsigned int number, std::vector<int>& out) const = 0;
+    virtual size_t query_nearest(CVecRef point, unsigned int number, std::vector<uint64_t>& out) const = 0;
 
     // Query for points within a certain radius from a point (range query)
-    virtual size_t query_sphere(CVecRef point, double radius, std::vector<int>& out) const = 0;
+    virtual size_t query_sphere(CVecRef point, double radius, std::vector<uint64_t>& out) const = 0;
 
     // Query for points in a box (range query)
-    virtual size_t query_box(CVecRef minCorner, CVecRef maxCorner, std::vector<int>& out) const = 0;
+    virtual size_t query_box(CVecRef minCorner, CVecRef maxCorner, std::vector<uint64_t>& out) const = 0;
 
     // Check if the index is empty
     // virtual bool is_empty() const = 0;
